fix: admins delete

+++ src/main/java/com/fredfonseca/bookstoremanager/users/service/UserService.java
@@ -1,16 +1,14 @@
 package com.fredfonseca.bookstoremanager.users.service;
 
 import com.fredfonseca.bookstoremanager.rentals.repository.RentalRepository;
-import com.fredfonseca.bookstoremanager.users.dto.AdminDTO;
-import com.fredfonseca.bookstoremanager.users.dto.MessageDTO;
-import com.fredfonseca.bookstoremanager.users.dto.UserDTO;
-import com.fredfonseca.bookstoremanager.users.dto.UserResponseDTO;
+import com.fredfonseca.bookstoremanager.users.dto.*;
 import com.fredfonseca.bookstoremanager.users.entity.Users;
 import com.fredfonseca.bookstoremanager.users.enums.Role;
 import com.fredfonseca.bookstoremanager.users.exception.*;
 import com.fredfonseca.bookstoremanager.users.mapper.UserMapper;
 import com.fredfonseca.bookstoremanager.users.repository.UserRepository;
 import com.fredfonseca.bookstoremanager.utils.StringPattern;
+import net.bytebuddy.implementation.bytecode.Throw;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -124,6 +122,22 @@ public class UserService {
     public void delete(Long id) {
         Users userToDelete = verifyAndGetIfExists(id);
 
+        if(userToDelete.getRole().getDescription().equals(ROLE_ADMIN)) {
+            throw new InvalidCredentialsChange(ROLE_USER);
+        }
+        if (!rentalRepository.findByUsers(userToDelete).isEmpty()) {
+            throw new DeleteDeniedException();
+        }
+        userRepository.deleteById(id);
+    }
+
+    public void deleteAdmin(Long id, AuthenticatedUser authenticatedUser) {
+        Users userToDelete = verifyAndGetIfExists(id);
+        Users foundAuthenticatedUser = verifyAndGetUserIfExists(authenticatedUser.getUsername());
+
+        if(userToDelete.getRole().getDescription().equals(ROLE_USER)) {
+            throw new InvalidCredentialsChange(ROLE_ADMIN);
+        }
         if(userToDelete.getUsername().equals("admin")) {
             throw new DeleteDeniedException(userToDelete.getUsername());
         }
